requirements.txt
telethon>=1.28.0
python-dotenv>=1.0.0
solana>=0.30.0
aiohttp>=3.8.0
solders>=0.18.0
base58>=2.1.0
flask>=2.3.0
flask-cors>=4.0.0




